etinker: consolidate toolchain's 'software.conf' information
'toolchain-{ref,url}' information is now pulled from the following:

etinker/boards/$(ET_CROSS_TUPLE)/software.conf

Previously the 'toolchain-{ref,url}' info was kept in each board's
software.conf file.

Additionally, pin 'toolchain-ref' to a crosstool-ng Git hash. This
is a slowly changing value in etinker.

+++ software.mk
@@ -19,45 +19,50 @@ define software-check
 	@if ! [ -d $(ET_SOFTWARE_DIR)/$1 ] || [ "fetch" = "$3" ]; then \
 		mkdir -p $(ET_SOFTWARE_DIR); \
 		(cd $(ET_SOFTWARE_DIR) && \
-			url="$(shell sed -n 's/$2-url://p' $(ET_BOARD_DIR)/software.conf)" && \
-			ref="$(shell sed -n 's/$2-ref://p' $(ET_BOARD_DIR)/software.conf)" && \
+			if [ "toolchain" = "$2" ]; then \
+				export et_url="$(shell sed -n 's/$2-url://p' $(ET_DIR)/boards/$(ET_CROSS_TUPLE)/software.conf)"; \
+				export et_ref="$(shell sed -n 's/$2-ref://p' $(ET_DIR)/boards/$(ET_CROSS_TUPLE)/software.conf)"; \
+			else \
+				export et_url="$(shell sed -n 's/$2-url://p' $(ET_BOARD_DIR)/software.conf)"; \
+				export et_ref="$(shell sed -n 's/$2-ref://p' $(ET_BOARD_DIR)/software.conf)"; \
+			fi; \
 			case $2 in \
 			toolchain | bootloader | kernel | rootfs) \
 				if [ -d $1 ]; then \
 					(cd $1 && git restore . && git clean -df && git fetch --all && git fetch --tags); \
 				else \
-					git clone $$url $1; \
+					git clone $$et_url $1; \
 				fi; \
 				if ! [ -d $1 ]; then \
 					printf "\n***** [$(ET_BOARD)][$(ET_BOARD_TYPE)] ERROR MISSING $(ET_SOFTWARE_DIR)/$1 DIRECTORY *****\n\n"; \
 					exit 2; \
 				fi; \
-				(cd $1 && git checkout $$ref) || exit 2; \
+				(cd $1 && git checkout $$et_ref) || exit 2; \
 				if [ -d $(ET_PATCH_DIR)/$(notdir $1) ]; then \
 					(cd $1 && \
 						git branch -D patched -f 2> /dev/null; \
 						git switch -c patched; \
 						patch -p1 -i $(ET_PATCH_DIR)/$(notdir $1)/*.patch && \
-						git commit -a -m "etinker: patches applied @ $$ref") || exit 2; \
+						git commit -a -m "etinker: patches applied @ $$et_ref") || exit 2; \
 				fi; \
 				;; \
 			*) \
 				if [ -d $2 ]; then \
 					(cd $2 && git restore . && git clean -df && git fetch --all && git fetch --tags); \
 				else \
-					git clone $$url $2; \
+					git clone $$et_url $2; \
 				fi; \
 				if ! [ -d $2 ]; then \
 					printf "\n***** [$(ET_BOARD)][$(ET_BOARD_TYPE)] ERROR MISSING $(ET_SOFTWARE_DIR)/$2 DIRECTORY *****\n\n"; \
 					exit 2; \
 				fi; \
-				(cd $2 && git checkout $$ref) || exit 2; \
+				(cd $2 && git checkout $$et_ref) || exit 2; \
 				if [ -d $(ET_PATCH_DIR)/$(notdir $2) ]; then \
 					(cd $2 && \
 						git branch -D patched -f 2> /dev/null; \
 						git switch -c patched; \
 						patch -p1 -i $(ET_PATCH_DIR)/$(notdir $2)/*.patch && \
-						git commit -a -m "etinker: patches applied @ $$ref") || exit 2; \
+						git commit -a -m "etinker: patches applied @ $$et_ref") || exit 2; \
 				fi; \
 				;; \
 			esac; \
